% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_tables.R
\name{append_tables}
\alias{append_tables}
\title{Database Appending Function}
\usage{
append_tables(
  db,
  tables,
  name,
  dt_coll = "date",
  colls = "*",
  return = TRUE,
  sqlite_temp = FALSE
)
}
\arguments{
\item{db}{connection to the database}

\item{tables}{vector of table present in the db to append}

\item{name}{new name of the resulting table}

\item{dt_coll}{name of the column that represents dates}

\item{colls}{which columns to append}

\item{return}{if TRUE, the function returns the resulting data frame}

\item{sqlite_temp}{if TRUE, the SQL table created is temporary}
}
\description{
Functions for appending tables into a new table in a database.
}
