% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{check_params}
\alias{check_params}
\title{Validate Input Parameters for CDM Data Processing}
\usage{
check_params(
  data_model,
  json_path_to_cdm_schema,
  format_source_files,
  folder_path_to_source_files,
  through_parquet,
  create_db_as
)
}
\arguments{
\item{data_model}{Character.
The name of the data model to use (e.g., `"conception"`).}

\item{json_path_to_cdm_schema}{Character.
Full path to the JSON file containing the CDM schema definition.}

\item{format_source_files}{Character.
Format of the source files. Must be either `"csv"` or `"parquet"`.}

\item{folder_path_to_source_files}{Character.
Path to the folder containing the source files.
Ensure the path ends with a trailing slash (`"/"`).}

\item{through_parquet}{Character.
Indicates whether to process through parquet files.
Must be either `"yes"` or `"no"`.}

\item{create_db_as}{Character.
Specifies whether to create the database using `"views"` or `"tables"`.}
}
\value{
No return value.
 The function stops execution with an error message if any check fails.
 If all checks pass, a confirmation message is printed.
}
\description{
This function performs a series of checks to validate the input parameters
required for processing source data into DuckDB.
}
\examples{
\dontrun{
check_params(
  data_model = "conception",
  json_path_to_cdm_schema = "schema/cdm_schema.json",
  format_source_files = "csv",
  folder_path_to_source_files = "data/source/",
  through_parquet = "no",
  create_db_as = "tables",
)
}

}
\keyword{internal}
