% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concePTION_metadata_v2}
\alias{concePTION_metadata_v2}
\title{Metadata for concePTION Common Data Model}
\format{
## `concePTION_metadata_v2`
A data frame with 263 rows and 6 columns:
\describe{
  \item{Variable}{variable name}
  \item{Mandatory}{indicator of whether this variable is mandatory}
  \item{Description}{description of the column with this variable}
  \item{Format}{format of the column}
  \item{TABLE}{name of the table in the database containing the variable}
  \item{cols_list}{column name for this variable}
}
}
\source{
<https://docs.google.com/spreadsheets/d/1hc-TBOfEzRBthGP78ZWIa13C0RdhU7bK/edit#gid=439480870> #nolint
}
\usage{
concePTION_metadata_v2
}
\description{
data.table describing the elements of the concePTION CDM database.
}
\keyword{datasets}
