% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indexes.R
\name{create_indexes}
\alias{create_indexes}
\title{Create Indexes for Tables in a Database}
\usage{
create_indexes(db_conn, specs)
}
\arguments{
\item{db_conn}{A database connection object.}

\item{specs}{A list containing the table names as keys and lists of column
names for indexing as values.}
}
\description{
This function creates indexes for tables in a database according to the
provided specifications.
}
\examples{
\dontrun{
# Example usage of create_indexes
db_conn <- dbConnect(duckdb::duckdb(), ":memory:")
specs <- list(
  persons = list("1" = "person_id", "2" = c("person_id", "sex_at_instance_creation")),
  medicines = list("1" = "person_id", "2" = c("person_id", "date_dispensing"))
)
create_indexes(db_conn, specs)
}
}
\keyword{internal}
