% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{ensure_data_table}
\alias{ensure_data_table}
\title{Ensure an object is a data.table.
It used the function data.table::as.data.table. This is very
permissive. Almost anything that can be converted to a data.table
by this function. Caution is advised.}
\usage{
ensure_data_table(
  obj,
  error_message = "Input must be convertible to a data.table"
)
}
\arguments{
\item{obj}{The object to check and convert.}

\item{error_message}{Custom error message if conversion fails.}
}
\value{
A data.table object.
}
\description{
Ensure an object is a data.table.
It used the function data.table::as.data.table. This is very
permissive. Almost anything that can be converted to a data.table
by this function. Caution is advised.
}
\keyword{internal}
