% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_origin_value.R
\name{get_origin_value}
\alias{get_origin_value}
\title{Get values from the origin database based on specified search_scheme}
\usage{
get_origin_value(cases_dt, db_connection, search_scheme = NULL)
}
\arguments{
\item{cases_dt}{Data table containing information about cases.
Must be a data.table with at least the search_scheme
"unique_id" and "ori_table".}

\item{db_connection}{Database connection object.
Must be an active DuckDB connection.}

\item{search_scheme}{A list specifying the search_scheme to retrieve}
}
\value{
A data table with the collected values.
}
\description{
This function retrieves values from the origin database for specified
search_scheme and tables.
}
\examples{
\dontrun{
get_origin_value(
  cases_dt,
  db_connection,
  list("Table1" = "Column1", "Table2" = "Column2")
)
}

}
