% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records_origin.R
\name{get_records_origin}
\alias{get_records_origin}
\title{Get Records from Origin CDM Tables within a DB Instance}
\usage{
get_records_origin(
  db_connection,
  unique_id_dt,
  unique_id_name = "ID",
  separator_id = "-"
)
}
\arguments{
\item{db_connection}{Database connection (SQLiteConnection).}

\item{unique_id_dt}{Data.table containing unique identifiers.}

\item{unique_id_name}{Name of the unique identifier. Default is "ID".}

\item{separator_id}{String that defines the separators between the table name
and the ROWID number.}
}
\value{
A list containing records from the origin CDM tables for each unique
  identifier.
}
\description{
This function returns the records from the origin CDM tables within a DB
instance that have a unique identifier.
}
\examples{
\dontrun{
# Example usage:
table <- dbReadTable(db_connection, "EVENTS")
cases <- as.data.table(table)[seq(1, 1000), "ID"]
a <- get_records_origin(db_connection = db_connection, cases,
  separator_id = "-")
}

}
