% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sql}
\alias{get_sql}
\title{Read SQL script from file and return as a single string
Function to make sure we read SQL scripts properly
(replace -- to /* .... */ top prevent line end errors)}
\usage{
get_sql(filepath)
}
\arguments{
\item{filepath}{The path to the SQL file}
}
\value{
A single string containing the SQL script
}
\description{
Read SQL script from file and return as a single string
Function to make sure we read SQL scripts properly
(replace -- to /* .... */ top prevent line end errors)
}
