% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{get_table_info}
\alias{get_table_info}
\title{Retrieve Column Names and Data Types for a Table}
\usage{
get_table_info(con, schema, table)
}
\arguments{
\item{schema}{Character.
The name of the schema containing the table.}

\item{table}{Character.
The name of the table for which to retrieve column metadata.}
}
\value{
A data frame with two columns: `column_name` and `data_type`.
}
\description{
Queries the DuckDB `information_schema.columns` table to retrieve
 the column names and data types for a specified table within a given schema.
}
\examples{
\dontrun{
get_table_info(schema = "cdm_conception", table = "person")
}

}
\keyword{internal}
