% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_codelist.R
\name{get_unique_codelist}
\alias{get_unique_codelist}
\title{Get Unique Code List from Database}
\usage{
get_unique_codelist(db_connection, scheme = NULL, column_info_list, tb_name)
}
\arguments{
\item{db_connection}{A database connection object.}

\item{scheme}{String defining the scheme where the table is hosted within the database. Default is NULL.}

\item{column_info_list}{A list of lists specifying the columns to be used for
distinct, along with their new names. Each inner list should contain two
elements: "source_column" and "alias_name".}

\item{tb_name}{Name of the CDM table.}
}
\value{
A list of data.tables with distinct values based
on the specified columns.
}
\description{
This function retrieves unique code lists from a database based on the
specified columns.
}
