% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_origin_dapspecific.R
\name{get_value_origin_dapspecific}
\alias{get_value_origin_dapspecific}
\title{Get values from the origin database based on DAP-specific codes}
\usage{
get_value_origin_dapspecific(cases_dt, db_connection, codelist)
}
\arguments{
\item{cases_dt}{Data table containing information about cases.}

\item{db_connection}{Database connection object.}

\item{codelist}{Data table containing DAP-specific code information.}
}
\value{
A data table with the collected values.
}
\description{
This function retrieves values from the origin database for specific 
DAP-specific codes.
}
\examples{
\dontrun{
# Example usage:
get_value_origin_dapspecific(cases_dt, db_connection, codelist)
}

}
