% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSQL}
\alias{getSQL}
\title{Read SQL script from file and return as a single string}
\usage{
getSQL(filepath)
}
\arguments{
\item{filepath}{The path to the SQL file}
}
\value{
A single string containing the SQL script
}
\description{
Read SQL script from file and return as a single string
}
