% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import and Format CSV Data}
\usage{
import_file(path, cols = NULL, cols_new = NULL, exprs = NULL, date_cols = NULL)
}
\arguments{
\item{path}{Path to the CSV file.}

\item{cols}{A vector of strings indicating which columns to import. Optional.}

\item{cols_new}{A vector of strings indicating the new column names.
Optional; must have the same length as cols.}

\item{exprs}{A data.table expression (within expression()) for subsetting
rows. Optional.}

\item{date_cols}{Vector of strings indicating variables with 'YYYYMMDD'
format to be converted to date format. Optional.}
}
\value{
A data.table data frame with formatted columns.
}
\description{
This function imports CSV files and applies basic formatting, such as
handling dates, handling missing values, and removing leading and trailing
spaces. It also allows for column selection and subsetting based on specified
criteria.
}
\examples{
\dontrun{
# Example usage of import_file
path <- "path/to/your/file.csv"
data <- import_file(path,
  cols = c("col1", "col2"),
  cols_new = c("new_col1", "new_col2"), exprs = expression(col1 > 0),
  date_cols = c("date_col")
)
}

}
\keyword{internal}
