% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_program.R
\name{import_program}
\alias{import_program}
\title{Import program}
\usage{
import_program(program_list, fl)
}
\arguments{
\item{program_list}{program list}

\item{fl}{file name}
}
\description{
Import meta data file based on directory and file name specified in the program_list
}
