% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_concept_table.R
\name{initialize_concept_table}
\alias{initialize_concept_table}
\title{Initialize Concept Table}
\usage{
initialize_concept_table(
  con,
  type_table = "view",
  path_parquets = NULL,
  partition = TRUE,
  overwrite = FALSE,
  add_id_set = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{type_table}{Character. Either `"view"` or `"table"`.}

\item{path_parquets}{Character. Path to Parquet files.}

\item{partition}{Logical. If `TRUE`, uses Hive partitioning.}
}
\value{
Invisibly returns the result of the `dbExecute` call or NULL if skipped.
}
\description{
Initialize Concept Table
}
