% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWEDataManagementModel-package.R
\docType{package}
\name{RWEDataManagementModel-package}
\alias{RWEDataManagementModel}
\alias{RWEDataManagementModel-package}
\title{RWEDataManagementModel: Data Management Model for Real World Evidence group}
\description{
RWE-DataManagementModel is an R package that provides a data model and functionality for processing RWD.
}
\author{
\strong{Maintainer}: Albert Cid Royo \email{A.CidRoyo@umcutrecht.nl}

Authors:
\itemize{
  \item Yinan Mao \email{Y.Mao@umcutrecht.nl}
  \item Roel Elbers
  \item Barbara Vreede \email{B.M.I.Vreede@umcutrecht.nl} (\href{https://orcid.org/0000-0002-5023-4601}{ORCID})
}

}
\keyword{internal}
