% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{setup_db_connection}
\alias{setup_db_connection}
\title{Set Up DuckDB Connection and Create Schemas}
\usage{
setup_db_connection(
  schema_individual_views,
  schema_conception,
  schema_combined_views,
  file_path_to_target_db
)
}
\arguments{
\item{schema_individual_views}{Character.
Name of the schema to store individual views.}

\item{schema_conception}{Character.
Name of the schema to store CDM tables.}

\item{schema_combined_views}{Character.
Name of the schema to store combined views.}

\item{file_path_to_target_db}{Character.
Full path to the DuckDB database file to be created.}
}
\value{
A DuckDB connection object (`DBIConnection`) with the schemas.
}
\description{
This function initializes a DuckDB connection and creates the specified
 schemas for organizing views and CDM tables. If a database file already
 exists at the specified path, it will be removed and replaced.
}
\examples{
\dontrun{
con <- setup_db_connection(
  schema_individual_views = "individual_views",
  schema_conception = "cdm_conception",
  schema_combined_views = "combined_views",
  file_path_to_target_db = "data/target/my_database.duckdb"
)
}
}
\keyword{internal}
