% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{suppress_cmd_checks}
\alias{suppress_cmd_checks}
\title{Suppress R CMD check notes for data.table's non-standard evaluation}
\description{
Suppress R CMD check notes for data.table's non-standard evaluation
}
\keyword{internal}
