% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2.DMM.R
\docType{package}
\name{T2.DMM-package}
\alias{T2.DMM}
\alias{T2.DMM-package}
\title{T2.DMM: Data Management Model for Real World Evidence}
\description{
RWE-DataManagementModel is an R package that provides a data model and functionality for processing RWD. T2 stands for the semantic harmonization transformation step.
}
\keyword{internal}
