% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicates_origin.R
\name{table_exists}
\alias{table_exists}
\title{Check if a table exists in a DuckDB database}
\usage{
table_exists(con, table_name, schema = NULL)
}
\arguments{
\item{con}{A DBI connection to a DuckDB database.}

\item{table_name}{Name of the table to check.}

\item{schema}{Optional. Name of the schema to check within.
Default is NULL, which checks all schemas.}
}
\value{
TRUE if the table exists, FALSE otherwise.
}
\description{
Check if a table exists in a DuckDB database
}
