% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{add_missing_tables_as_empty}
\alias{add_missing_tables_as_empty}
\title{Add Missing Tables as Empty Views or Tables}
\usage{
add_missing_tables_as_empty(
  con,
  data_model,
  tables_in_cdm,
  files_in_input,
  create_db_as
)
}
\arguments{
\item{con}{A DBI connection object to the target database.}

\item{data_model}{String.
The name of the data model (e.g., `"conception"`).}

\item{tables_in_cdm}{Character vector.
List of expected table names in the CDM.}

\item{files_in_input}{Character.
Named character vector of target and source files.}

\item{create_db_as}{Character.
Whether to create `"views"` or `"tables"` in the combined schema.}
}
\value{
No return value. The function performs side effects by executing
 SQL statements to create views or tables.
}
\description{
This function checks for missing tables in the input files compared to the
 expected tables in the CDM, and creates empty views or tables for them in
 the database.
}
\examples{
\dontrun{
add_missing_tables_as_empty(
  con = db_connection,
  data_model = "my_model",
  tables_in_cdm = c("person", "visit_occurrence"),
  files_in_input = list(person = "person.csv"),
  create_db_as = "views"
)
}

}
\keyword{internal}
