% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{check_params}
\alias{check_params}
\title{Validate Input Parameters for CDM Data Processing}
\usage{
check_params(
  data_model,
  excel_path_to_cdm_schema,
  format_source_files,
  folder_path_to_source_files,
  through_parquet,
  file_path_to_target_db,
  create_db_as,
  verbosity
)
}
\arguments{
\item{data_model}{Character.
The name of the data model to use (e.g., `"conception"`).}

\item{excel_path_to_cdm_schema}{Character.
Full path to the Excel file containing the CDM schema definition.}

\item{format_source_files}{Character.
Format of the source files. Must be either `"csv"` or `"parquet"`.}

\item{folder_path_to_source_files}{Character.
Path to the folder containing the source files.
Ensure the path ends with a trailing slash (`"/"`).}

\item{through_parquet}{Character.
Indicates whether to process through parquet files.
Must be either `"yes"` or `"no"`.}

\item{file_path_to_target_db}{Character.
Full path to the target DuckDB database file to be created.}

\item{create_db_as}{Character.
Specifies whether to create the database using `"views"` or `"tables"`.}

\item{verbosity}{Integer.
Level of verbosity for logging.
Use `0` for minimal output and `1` for detailed output.}
}
\value{
No return value.
 The function stops execution with an error message if any check fails.
 If all checks pass, a confirmation message is printed.
}
\description{
This function performs a series of checks to validate the input parameters
required for processing source data into DuckDB.
}
\details{
The function performs the following checks:
\itemize{
  \item Validates existence of the source folder and schema file.
  \item Ensures the target database file does not already exist.
  \item Checks for validity of source files (`.csv` or `.parquet`).
  \item Validates the `data_model`, `create_db_as`, `verbosity`,
        and `through_parquet` parameters.
  \item Attempts to connect to the DuckDB database.
}
}
\examples{
\dontrun{
check_params(
  data_model = "conception",
  excel_path_to_cdm_schema = "schema/cdm_schema.xlsx",
  format_source_files = "csv",
  folder_path_to_source_files = "data/source/",
  through_parquet = "no",
  file_path_to_target_db = "data/target/my_database.duckdb",
  create_db_as = "tables",
  verbosity = 1
)
}

}
\keyword{internal}
