% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dap_specific_codelist.R
\name{create_dap_specific_codelist}
\alias{create_dap_specific_codelist}
\title{Create DAP-Specific Codelist}
\usage{
create_dap_specific_codelist(
  dap_codes,
  codelist,
  start_with_codingsystems = c("ICD10CM", "ICD10", "ICD10DA", "ICD9CM", "MTHICD9",
    "ICPC", "ICPC2P", "ICPC2EENG", "ATC"),
  priority_col = "priority"
)
}
\arguments{
\item{dap_codes}{Data.table containing unique code list output.}

\item{codelist}{Data.table containing the study code list.}

\item{start_with_codingsystems}{Columns to start with.}

\item{priority_col}{Name of the priority column to use for sorting. Defaults to "priority".}
}
\description{
Create DAP-Specific Codelist
}
