% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dap_specific_concept.R
\name{create_dap_specific_concept}
\alias{create_dap_specific_concept}
\title{Create DAP-Specific Concepts into Database}
\usage{
create_dap_specific_concept(
  codelist,
  name_attachment,
  save_db,
  date_col_filter = NULL,
  date_col_notnull = FALSE,
  dir_save = NULL,
  add_meaning = FALSE,
  save_in_parquet = FALSE,
  table_name = "cdm_table_name",
  column_name_prefix = "column_name",
  expected_value_prefix = "expected_value",
  keep_date_prefix = "keep_date",
  keep_column_prefix = "keep_value",
  intermediate_type = "TABLE",
  partition_var = "concept_id",
  add_tag = FALSE
)
}
\arguments{
\item{codelist}{A data.table containing information about tables,
columns, and values for DAP-specific concepts.}

\item{name_attachment}{Name of the table to be insert into the database.}

\item{save_db}{The database connection object where the edited tables
and concepts will be saved.}

\item{date_col_filter}{An optional filter to subset data based on
a specified date column.}

\item{date_col_notnull}{Logical; if TRUE and a date column is available,
keeps only rows where the selected date column is not NULL.}

\item{dir_save}{Directory path for parquet output (required if save_in_parquet = TRUE)}

\item{add_meaning}{A boolean that defines whether to save the meaning of any
CDM table -if available- in the results of the function.
This is specific for the ConcePTION CDM. Default: FALSE.]}

\item{save_in_parquet}{Logical; if TRUE, exports to parquet;
if FALSE, inserts to concept_table}

\item{table_name}{Name of the table in the codelist database.
Default: "cdm_table_name".}

\item{column_name_prefix}{An optional string that defines the prefix name
of the column name variable column(s) from the DAP-specific concept map.
Default: "column_name".}

\item{expected_value_prefix}{An optional string that defines the prefix name
of the expected value variable column(s) from the DAP-specific concept map.
Default: "expected_value".}

\item{keep_date_prefix}{Characther default: keep_date.
The prefix value to identify the column where the column name with the
date will be stored in the concept_table.}

\item{keep_column_prefix}{Characther default: keep_value
The prefix value to identify the column where the column name with the value
will stored in the concept_table.}

\item{intermediate_type}{Type of intermediate structure to create.}

\item{partition_var}{Default: concept_id.
Concept_table column to partition on.}

\item{add_tag}{Boolean defining whether we hardcode a tag = 1 or not. Default FALSE}
}
\description{
This function Creates DAP-specific concepts into a database using a
DAP-specifc concept map (codelist).
The function modifies the specified save_db by creating edited tables
and Creating DAP-specific concepts.
}
