% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dap_specific_concept.R
\name{create_dap_specific_concept}
\alias{create_dap_specific_concept}
\title{Create DAP-Specific Concepts into Database}
\usage{
create_dap_specific_concept(
  codelist,
  data_db,
  name_attachment,
  save_db,
  date_col_filter = NULL,
  table_name = "cdm_table_name",
  column_name_prefix = "column_name",
  expected_value_prefix = "expected_value",
  add_meaning = FALSE
)
}
\arguments{
\item{codelist}{A data.table containing information about tables, columns, and values for DAP-specific concepts.}

\item{data_db}{The database connection object.}

\item{name_attachment}{Attachment to the database table names.}

\item{save_db}{The database connection object where the edited tables and concepts will be saved.}

\item{date_col_filter}{An optional filter to subset data based on a specified date column.}

\item{table_name}{Name of the table in the codelist}

\item{column_name_prefix}{An optional string that defines the prefix name of the column name variable column(s) from the DAP-specific concept map}

\item{expected_value_prefix}{An optional string that defines the prefix name of the expected value variable column(s) from the DAP-specific concept map}

\item{add_meaning}{An optional boolean that defines whether the possibility to save the meaning of any CDM table -if available- in the results of the function. This is specific for the ConcePTION CDM.}
}
\description{
This function Creates DAP-specific concepts into a database using a DAP-specifc concept map (codelist).
The function modifies the specified save_db by creating edited tables and Creating DAP-specific concepts.
}
\examples{
\dontrun{
# Example usage of Create_dap_specific_concept
codelist <- data.table::fread("path/to/codelist.csv")
data_db <- dbConnect(RSQLite::SQLite(), ":memory:")
save_db <- dbConnect(RSQLite::SQLite(), ":memory:")
create_dap_specific_concept(codelist, data_db, "Attachment", save_db, date_col_filter = "20230101")
}

}
