% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{create_schemas}
\alias{create_schemas}
\title{Set Up DuckDB Connection and Create Schemas}
\usage{
create_schemas(schema_individual_views, data_model, con)
}
\arguments{
\item{schema_individual_views}{Character.
Name of the schema to store individual views.}

\item{data_model}{Character.
Name of the schema to store tables.}

\item{con}{DBIConnection.
The function will create schemas in this database.}
}
\description{
This function initializes a DuckDB connection and creates the specified
 schemas for organizing views and CDM tables. If a database file already
 exists at the specified path, it will be removed and replaced.
}
\examples{
\dontrun{
con <- setup_db_connection(
  schema_individual_views = "individual_views",
  data_model = NULL,
  file_path_to_target_db = "data/target/my_database.duckdb"
)
}
}
\keyword{internal}
