% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLoader.R
\docType{class}
\name{DatabaseLoader}
\alias{DatabaseLoader}
\title{DatabaseLoader Class}
\description{
The `DatabaseLoader` class manages the database connection,
loading configuration and metadata, and executing database operations in a
specified order.
}
\details{
This class provides methods to initialize a database connection, load the
database with required data, and execute a series of operations defined in
external R class scripts.
The operations are executed in the order specified in the configuration file.
The class also includes utility methods for cleaning up temporary files
and retrieving all operations from a specified folder.
}
\section{Methods}{

\describe{
  \item{`initialize(db_path, config_path, cdm_metadata)`}{
    Initializes the class with the provided parameters.
    \itemize{
      \item `db_path`: A string representing the path to the database file.
      \item `config_path`: A string representing the path to configuration.
      \item `cdm_metadata`: A string representing the path to the metadata.
    }
  }
  \item{`set_database()`}{
    Loads the database with the required data.
  }
  \item{`run_db_ops(ops)`}{
    Executes a series of database operations in the order of config file.
    \itemize{
      \item `ops`: A list of operation objects to execute. If `NULL, they
      are retrieved using the `get_all_operations` method.
    }
  }
}
}

\examples{
\dontrun{
loader <- DatabaseLoader$new(
  db_path = "path/to/database.db",
  config_path = "path/to/set_db.json",
  cdm_metadata = "path/to/CDM_metadata.rds"
)
loader$set_database()
loader$run_db_ops()
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db_path}}{A string representing the path to the database file.}

\item{\code{data_instance}}{A string representing the path to the data instance.}

\item{\code{config_path}}{A string representing the path to the configuration file.}

\item{\code{cdm_metadata}}{A string representing the path to the metadata file.}

\item{\code{config}}{A list containing the configuration from `config_path` file.}

\item{\code{metadata}}{A `data.table` object containing the metadata loaded.}

\item{\code{db}}{A database connection object created using the `duckdb` package.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseLoader-new}{\code{DatabaseLoader$new()}}
\item \href{#method-DatabaseLoader-set_database}{\code{DatabaseLoader$set_database()}}
\item \href{#method-DatabaseLoader-run_db_ops}{\code{DatabaseLoader$run_db_ops()}}
\item \href{#method-DatabaseLoader-clone}{\code{DatabaseLoader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseLoader-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseLoader-new}{}}}
\subsection{Method \code{new()}}{
Initializes the class with the provided parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseLoader$new(
  db_path = NULL,
  data_instance = NULL,
  config_path = NULL,
  cdm_metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{A string representing the path to the database file.}

\item{\code{data_instance}}{A string representing the path to the data instance.}

\item{\code{config_path}}{A string representing the path to configuration file.}

\item{\code{cdm_metadata}}{A string representing the path to the metadata file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseLoader-set_database"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseLoader-set_database}{}}}
\subsection{Method \code{set_database()}}{
Loads the database with the required data using config and metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseLoader$set_database()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseLoader-run_db_ops"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseLoader-run_db_ops}{}}}
\subsection{Method \code{run_db_ops()}}{
Executes a series of database operations in order specified in config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseLoader$run_db_ops(ops = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ops}}{A list of operation objects to execute. If `NULL`, operations
are retrieved using the `get_all_operations` method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseLoader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseLoader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseLoader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
