% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplicateRemover.R
\docType{class}
\name{DuplicateRemover}
\alias{DuplicateRemover}
\title{DuplicateRemover Class}
\description{
The `DuplicateRemover` class is a subclass of the `DatabaseOperation` class.
It is responsible for identifying and removing duplicate records from the
database tables specified in the `db_loader` object.
}
\details{
This class uses the `delete_duplicates_origin` function from the `T2.DMM`
package to remove duplicates. The duplicates are removed based on the
schema provided by the `db_loader` object. Optionally, the deleted records
can be saved to an intermediate file for auditing or debugging purposes.
}
\section{Methods}{

\describe{
  \item{`run(db_loader)`}{
    Executes the duplicate removal process.
    \itemize{
      \item `db_loader`: A `DatabaseLoader` object provides db connection,
      table names, and configuration details required for the operation.
    }
  }
}
}

\keyword{internal}
\section{Super class}{
\code{\link[T2.DMM:DatabaseOperation]{T2.DMM::DatabaseOperation}} -> \code{DuplicateRemover}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{A string representing the name of the class.
Default is `"DuplicateRemover"`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuplicateRemover-run}{\code{DuplicateRemover$run()}}
\item \href{#method-DuplicateRemover-clone}{\code{DuplicateRemover$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="T2.DMM" data-topic="DatabaseOperation" data-id="is_enabled"><a href='../../T2.DMM/html/DatabaseOperation.html#method-DatabaseOperation-is_enabled'><code>T2.DMM::DatabaseOperation$is_enabled()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRemover-run"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRemover-run}{}}}
\subsection{Method \code{run()}}{
Executes the duplicate removal process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRemover$run(db_loader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_loader}}{A `DatabaseLoader` object provides database connection,
table names, and other configuration details required for the operation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRemover-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRemover-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRemover$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
