% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSTReportGenerator.R
\name{FSTReportGenerator}
\alias{FSTReportGenerator}
\title{FSTReportGenerator Class}
\description{
The `FSTReportGenerator` class is a specialized subclass of `ReportGenerator`
designed to generate and save tabular reports in the `.fst` file format.
It uses the `fst::write_fst()` function to persist data efficiently.

This class is automatically selected when the report extension is `.fst`,
as determined by the `ReportGenerator`'s internal logic. The generated report
is saved in the directory and with the filename specified in the
`report` configuration section of the provided `DatabaseLoader` object.
}
\details{
The main method `write_report()` takes the output of a table row count query
and writes it to a `.fst` file. The class also exposes a `supported_ext()`
method, which declares the supported file extension(s) for dynamic dispatch.
}
\section{Methods}{

\describe{
  \item{`write_report(data, db_loader)`}{
    Writes the provided `data` to a `.fst`.
  }
  \item{`supported_ext()`}{
    Returns a character vector indicating file extensions (here, `"fst"`).
  }
}
}

\keyword{internal}
\section{Super classes}{
\code{\link[T2.DMM:DatabaseOperation]{T2.DMM::DatabaseOperation}} -> \code{\link[T2.DMM:ReportGenerator]{T2.DMM::ReportGenerator}} -> \code{FSTReportGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{A string identifying the class name.
Default is `"FSTReportGenerator"`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FSTReportGenerator-write_report}{\code{FSTReportGenerator$write_report()}}
\item \href{#method-FSTReportGenerator-supported_ext}{\code{FSTReportGenerator$supported_ext()}}
\item \href{#method-FSTReportGenerator-clone}{\code{FSTReportGenerator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="T2.DMM" data-topic="DatabaseOperation" data-id="is_enabled"><a href='../../T2.DMM/html/DatabaseOperation.html#method-DatabaseOperation-is_enabled'><code>T2.DMM::DatabaseOperation$is_enabled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="T2.DMM" data-topic="ReportGenerator" data-id="run"><a href='../../T2.DMM/html/ReportGenerator.html#method-ReportGenerator-run'><code>T2.DMM::ReportGenerator$run()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTReportGenerator-write_report"></a>}}
\if{latex}{\out{\hypertarget{method-FSTReportGenerator-write_report}{}}}
\subsection{Method \code{write_report()}}{
Writes the provided data to a `fst` file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSTReportGenerator$write_report(data, db_loader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing the data to be saved.}

\item{\code{db_loader}}{A `DatabaseLoader` object that provides the
configuration details for saving the report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTReportGenerator-supported_ext"></a>}}
\if{latex}{\out{\hypertarget{method-FSTReportGenerator-supported_ext}{}}}
\subsection{Method \code{supported_ext()}}{
Returns a character vector indicating file extensions (here, `"fst"`).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSTReportGenerator$supported_ext()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTReportGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FSTReportGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSTReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
