% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{generate_ddl}
\alias{generate_ddl}
\title{Generate DDL (Data Definition Language) for a DuckDB Table}
\usage{
generate_ddl(df, data_model, table_name, schema_name)
}
\arguments{
\item{df}{Data Frame.
Must contain `Variable` and `Format` columns.}

\item{data_model}{String.}

\item{table_name}{String.}

\item{schema_name}{String.}
}
\value{
A character string containing the full DDL SQL statement.
}
\description{
This function creates a SQL `CREATE OR REPLACE TABLE` statement for a given
 table based on a data frame containing variable names and formats. It maps
 formats to DuckDB-compatible data types and constructs the full DDL string.
}
\keyword{internal}
