% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_origin_value.R
\name{get_origin_value}
\alias{get_origin_value}
\title{Get values from the origin database based on specified columns}
\usage{
get_origin_value(cases_dt, db_connection, search_scheme = NULL)
}
\arguments{
\item{cases_dt}{Data table containing information about cases.}

\item{db_connection}{Database connection object.}

\item{search_scheme}{A list specifying the columns to retrieve for each unique
ori_table.}
}
\value{
A data table with the collected values.
}
\description{
This function retrieves values from the origin database for specified columns
and tables.
}
\examples{
\dontrun{
# Example usage:
get_origin_value(
  cases_dt, db_connection,
  list("Table1" = "Column1", "Table2" = "Column2")
)
}

}
