% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_codelist.R
\name{get_unique_codelist}
\alias{get_unique_codelist}
\title{Get Unique Code List from Database}
\usage{
get_unique_codelist(db_connection, column_info_list, tb_name)
}
\arguments{
\item{db_connection}{A database connection object.}

\item{column_specs}{A list of lists specifying the columns to be used for
distinct, along with their new names. Each inner list should contain two
elements: "column_name" and "new_name".}
}
\value{
A data.table with distinct values based on the specified columns.
}
\description{
This function retrieves unique code lists from a database based on the
specified columns.
}
\examples{
\dontrun{
# Example usage:
# Define the column information list
column_info_list <- list(
  list(column_name = "event_code", alias_name = "code"),
  list(column_name = "event_record_vocabulary", alias_name = "coding_system"),
  list(column_name = "event_free_text", alias_name = "code")
)
 column_info_list <- list(
list(column_name = c("event_code", "event_record_vocabulary"),
     alias_name = c("code", "coding_system")),
list(column_name = c("event_free_text"),
     alias_name = c("code"))
)

# Replace 'your_db_connection' with your actual database connection
db_connection_origin <- DBI::dbConnect(RSQLite::SQLite(), dbname = "your_database.db")

# Call the function with the database connection and column information list
result_list <- get_unique_codelist(db_connection_origin, column_info_list)

# Close the database connection
DBI::dbDisconnect(db_connection_origin)

# Access individual results from the list
unique_codelist_event_code <- result_list[[1]]
unique_codelist_free_text <- result_list[[2]]
}

}
