% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dap_specific_codelist.R
\name{import_dap_specific_codelist}
\alias{import_dap_specific_codelist}
\title{Import DAP-Specific Codelist to Database}
\usage{
import_dap_specific_codelist(
  codelist_path,
  codelist_name_db,
  db_connection,
  columns,
  ...
)
}
\arguments{
\item{codelist_path}{String. Path to the .rds file.}

\item{codelist_name_db}{String. Target table name in the database.}

\item{db_connection}{DBI connection object.}

\item{columns}{Character vector. Columns to keep and transform.}
}
\description{
Reads a codelist, filters for active rows ("BOTH"), converts specified 
columns to uppercase, and writes the result to a database.
}
