% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db_functions.R
\name{populate_cdm_tables_from_views}
\alias{populate_cdm_tables_from_views}
\title{Populate CDM Tables from Source Views}
\usage{
populate_cdm_tables_from_views(
  db_connection,
  data_model,
  schema_individual_views,
  schema_conception,
  files_in_input,
  through_parquet,
  create_db_as
)
}
\arguments{
\item{db_connection}{A DuckDB database connection object (`DBIConnection`).}

\item{data_model}{Character.
The name of the data model (e.g., `"conception"`).}

\item{schema_individual_views}{Character.
Schema containing the source views.}

\item{schema_conception}{Character.
Schema where the CDM tables are located.}

\item{files_in_input}{Named character vector.
Sanitized view names as values, with CDM table names as names.}

\item{through_parquet}{Character.
Whether to use Parquet as an intermediate step (`"yes"` or `"no"`).}

\item{create_db_as}{Character.
Indicates whether the database was created using `"views"` or `"tables"`.}

\item{file_path_to_target_db}{Character.
Full path to the DuckDB database file.}
}
\value{
No return value.
 The function inserts data into CDM tables and logs progress to `log.txt`.
}
\description{
This function populates CDM tables in DuckDB by transforming and inserting
data from previously created source views.
}
\examples{
\dontrun{
populate_cdm_tables_from_views(
  db_connection = con,
  data_model = "conception",
  schema_individual_views = "individual_views",
  schema_conception = "cdm_conception",
  files_in_input = c(person1 = "person", person2 = "person"),
  through_parquet = "no",
  file_path_to_target_db = "data/target/my_database.duckdb",
  create_db_as = "tables"
)
}

}
\keyword{internal}
