% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportGenerator.R
\docType{class}
\name{ReportGenerator}
\alias{ReportGenerator}
\title{ReportGenerator Class}
\description{
The `ReportGenerator` class is a subclass of the `DatabaseOperation` class.
It is responsible for generating reports based on the data in the database.
}
\details{
This class generates a report containing the row counts for all tables in the
database. The row counts are retrieved using the `get_rows_tables` function
from the `T2.DMM` package. The report is saved as a `.fst` file in the
directory specified in the `report` configuration of the `db_loader` object.
The database connection is closed after the report is generated.
}
\section{Methods}{

\describe{
  \item{`run(db_loader)`}{
    Executes the report generation process.
    \itemize{
      \item `db_loader`: A `DatabaseLoader` object provides db connection,
      the instance name, and the configuration details for saving the report.
    }
  }
}
}

\examples{
\dontrun{
loader <- DatabaseLoader$new()
report_generator <- ReportGenerator$new()
report_generator$run(loader)
}

}
\keyword{internal}
\section{Super class}{
\code{\link[T2.DMM:DatabaseOperation]{T2.DMM::DatabaseOperation}} -> \code{ReportGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{A string representing the name of the class.
Default is `"ReportGenerator"`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReportGenerator-run}{\code{ReportGenerator$run()}}
\item \href{#method-ReportGenerator-clone}{\code{ReportGenerator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="T2.DMM" data-topic="DatabaseOperation" data-id="is_enabled"><a href='../../T2.DMM/html/DatabaseOperation.html#method-DatabaseOperation-is_enabled'><code>T2.DMM::DatabaseOperation$is_enabled()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-ReportGenerator-run}{}}}
\subsection{Method \code{run()}}{
Executes the report generation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$run(db_loader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_loader}}{A `DatabaseLoader` object provides database connection,
the instance name, and the configuration details for saving the report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReportGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
