% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2.DMM.R
\docType{package}
\name{T2.DMM-package}
\alias{T2.DMM}
\alias{T2.DMM-package}
\title{T2.DMM: Data Management Model for Real World Evidence}
\description{
RWE-DataManagementModel is an R package that provides a data model and functionality for processing RWD. T2 stands for the semantic harmonization transformation step.
}
\author{
\strong{Maintainer}: Albert Cid Royo \email{A.CidRoyo@umcutrecht.nl} (\href{https://orcid.org/0000-0002-1863-3450}{ORCID})

Authors:
\itemize{
  \item Yinan Mao \email{Y.Mao@umcutrecht.nl}
  \item Barbara Vreede \email{B.M.I.Vreede@umcutrecht.nl} (\href{https://orcid.org/0000-0002-5023-4601}{ORCID})
}

}
\keyword{internal}
