% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueIdGenerator.R
\docType{class}
\name{UniqueIdGenerator}
\alias{UniqueIdGenerator}
\title{UniqueIdGenerator Class}
\description{
The `UniqueIdGenerator` class is a subclass of the `DatabaseOperation` class.
It is responsible for generating unique identifiers for the records in the
specified database tables.
}
\details{
This class uses the `create_unique_id` function from the `T2.DMM` package to
generate unique IDs for the tables specified in the `cdm_tables_names`
configuration of the `db_loader` object. The unique IDs are generated
based on the instance name provided in the `db_loader` object.
}
\section{Fields}{

\describe{
  \item{`classname`}{
    A string representing the name of the class.
    Default is `"UniqueIdGenerator"`.
  }
}
}

\section{Methods}{

\describe{
  \item{`run(db_loader)`}{
    Executes the unique ID generation process.
    \itemize{
      \item `db_loader`: A `DatabaseLoader` object that provides the
       database connection, the list of tables to process, and the
       instance name used for generating unique IDs.
    }
  }
}
}

\examples{
\dontrun{
loader <- DatabaseLoader$new()
unique_id_generator <- UniqueIdGenerator$new()
unique_id_generator$run(loader)
}

}
\keyword{internal}
\section{Super class}{
\code{T2.DMM::DatabaseOperation} -> \code{UniqueIdGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{A string representing the name of the class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UniqueIdGenerator-run}{\code{UniqueIdGenerator$run()}}
\item \href{#method-UniqueIdGenerator-clone}{\code{UniqueIdGenerator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="T2.DMM" data-topic="DatabaseOperation" data-id="is_enabled"><a href='../../T2.DMM/html/DatabaseOperation.html#method-DatabaseOperation-is_enabled'><code>T2.DMM::DatabaseOperation$is_enabled()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UniqueIdGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-UniqueIdGenerator-run}{}}}
\subsection{Method \code{run()}}{
Executes the unique ID generation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniqueIdGenerator$run(db_loader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_loader}}{A `DatabaseLoader` object provides database connection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UniqueIdGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UniqueIdGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniqueIdGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
